requirements.txt
torch=1.10.0
torchvision=0.8.2
kornia=0.6.2
pillow=8.1.2
opencv-python=4.5.4.58